expense

+++ Delta/database/migrations/2020_09_24_170425_create_order_details_table.php
@@ -18,12 +18,12 @@ class CreateOrderDetailsTable extends Migration
 
             $table->unsignedBigInteger('order_id');
             $table->unsignedBigInteger('product_id');
-            $table->double('price',8,2)->default(0);
-            $table->double('quantity',8,6)->default(1);
-            $table->double('discount',8,2)->default(0);
-            $table->double('tax',8,2)->default(0);
-            $table->double('total',8,2)->default(0);
-            $table->bigInteger('profit')->default(0);
+            $table->double('price',18,2)->default(0);
+            $table->double('quantity',18,6)->default(1);
+            $table->double('discount',18,2)->default(0);
+            $table->double('tax',18,2)->default(0);
+            $table->double('total',18,2)->default(0);
+            $table->double('profit',18,2)->default(0);
 
             $table->softDeletes();
